+++ src/libponyrt/lang/posix_except.c
@@ -7,12 +7,14 @@
 #include <unwind.h>
 #include <stdlib.h>
 
-PONY_EXTERN_C_BEGIN
-
 #ifdef PLATFORM_IS_ARM
 #include <string.h>
-#include <stdio.h>
+#define PONY_EXCEPTION_CLASS "Pony\0\0\0\0"
+#else
+#define PONY_EXCEPTION_CLASS 0x506F6E7900000000 // "Pony"
 #endif
+
+PONY_EXTERN_C_BEGIN
 
 static __pony_thread_local struct _Unwind_Exception exception;
 static __pony_thread_local uintptr_t landing_pad;
@@ -27,9 +29,9 @@ static void exception_cleanup(_Unwind_Reason_Code reason,
 PONY_API void pony_error()
 {
 #ifdef PLATFORM_IS_ARM
-  memcpy(exception.exception_class, "Pony\0\0\0\0", 8);
+  memcpy(exception.exception_class, PONY_EXCEPTION_CLASS, 8);
 #else
-  exception.exception_class = 0x506F6E7900000000; // "Pony"
+  exception.exception_class = PONY_EXCEPTION_CLASS;
 #endif
   exception.exception_cleanup = exception_cleanup;
   _Unwind_RaiseException(&exception);
@@ -59,11 +61,14 @@ static _Unwind_Reason_Code continue_unwind(_Unwind_Exception* exception,
   return _URC_CONTINUE_UNWIND;
 }
 
-PONY_API _Unwind_Reason_Code pony_personality_v0(_Unwind_State state,
+PONY_API _Unwind_Reason_Code ponyint_personality_v0(_Unwind_State state,
   _Unwind_Exception* exception, _Unwind_Context* context)
 {
   if(exception == NULL || context == NULL)
     return _URC_FAILURE;
+
+  if(memcmp(exception->exception_class, PONY_EXCEPTION_CLASS, 8) != 0)
+    return continue_unwind(exception, context);
 
   // Save exception in r12.
   _Unwind_SetGR(context, 12, (uintptr_t)exception);
@@ -116,14 +121,15 @@ PONY_API _Unwind_Reason_Code pony_personality_v0(_Unwind_State state,
 
 #else
 
-PONY_API _Unwind_Reason_Code pony_personality_v0(int version,
+PONY_API _Unwind_Reason_Code ponyint_personality_v0(int version,
   _Unwind_Action actions, uint64_t ex_class,
   struct _Unwind_Exception* exception, struct _Unwind_Context* context)
 {
-  (void)ex_class;
-
   if(version != 1 || exception == NULL || context == NULL)
     return _URC_FATAL_PHASE1_ERROR;
+
+  if(ex_class != PONY_EXCEPTION_CLASS)
+    return _URC_CONTINUE_UNWIND;
 
   // The search phase sets up the landing pad.
   if(actions & _UA_SEARCH_PHASE)